move `expand_title_template` to fingerprinting utils module
