+++ src/sentry/grouping/utils.py
@@ -1,14 +1,11 @@
 from __future__ import annotations
 
-from collections.abc import Iterable, Mapping
+from collections.abc import Iterable
 from hashlib import md5
-from re import Match
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 from uuid import UUID
 
 from django.utils.encoding import force_bytes
-
-from sentry.grouping.fingerprinting.utils import _fingerprint_var_re, resolve_fingerprint_variable
 
 if TYPE_CHECKING:
     from sentry.grouping.component import ExceptionGroupingComponent
@@ -39,25 +36,3 @@ def bool_from_string(value: str) -> bool | None:
             return False
 
     return None
-
-
-def expand_title_template(
-    template: str, event_data: Mapping[str, Any], use_legacy_unknown_variable_handling: bool = False
-) -> str:
-    def _handle_match(match: Match[str]) -> str:
-        variable_key = match.group(1)
-        # TODO: Once we have fully transitioned off of the `newstyle:2023-01-11` grouping config, we
-        # can remove `use_legacy_unknown_variable_handling` and just return the value given by
-        # `resolve_fingerprint_variable`
-        resolved_value = resolve_fingerprint_variable(
-            variable_key, event_data, use_legacy_unknown_variable_handling
-        )
-
-        # TODO: Once we have fully transitioned off of the `newstyle:2023-01-11` grouping config, we
-        # can remove this
-        if resolved_value is not None:
-            return resolved_value
-        # If the variable can't be resolved, return it as is
-        return match.group(0)
-
-    return _fingerprint_var_re.sub(_handle_match, template)
